release 10.35.28
git-svn-id: http://svn.code.sf.net/p/netpbm/code/stable@328 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/other/pamtofits.c
@@ -32,8 +32,8 @@
 struct cmdlineInfo {
     const char * inputFileName;
     unsigned int maxSpec;
-    double max;
-    double min;
+    float max;
+    float min;
 };
